Moved Extensions.kt to util package
